Refactor navmesh into new module
